requirements.txt
numpy==1.19.3
tensorflow-cpu
Keras
Flask
gunicorn
Pillow
